+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ca.dart
@@ -607,4 +607,13 @@ class StreamChatLocalizationsCa extends GlobalStreamChatLocalizations {
 
   @override
   String get loadingPollVotesError => 'Error en carregar els vots';
+
+  @override
+  String get repliedToLabel => 'resposta a:';
+
+  @override
+  String newThreadsLabel({required int count}) {
+    if (count == 1) return '1 fil nou';
+    return '$count fils nous';
+  }
 }